feat: [김원기] Google OAuth 구현 [IFA-SPRING-16]

+++ src/main/java/com/eddicorp/ifaspring/user/repository/UserRepository.java
@@ -4,6 +4,7 @@ import com.eddicorp.ifaspring.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
@@ -11,7 +12,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByNickName(String nickName);
 
     @Query("select a from User a where a.oauthId=:kakaoId and a.oauthPlatform=:platformName ")
-    Optional<User> findByOauthId(Long kakaoId, String platformName);
+    Optional<User> findByOauthId(BigInteger kakaoId, String platformName);
     Boolean existsUserByNickName(String nickname);
     Boolean existsUserByEmail(String email);
