+++ src/main/java/com/eddicorp/ifaspring/user/service/UserService.java
@@ -5,11 +5,13 @@ import com.eddicorp.ifaspring.user.controller.form.LoginResForm;
 import com.eddicorp.ifaspring.user.entity.User;
 import org.springframework.http.HttpHeaders;
 
+import java.math.BigInteger;
+
 public interface UserService {
 
     User findByUserToken(String userToken);
-    LoginResForm loginOauthUser(Long platformId, String platformName, String profileImage);
-    User joinOauthUser(Long platformId, String platformName, String profileImage);
+    LoginResForm loginOauthUser(BigInteger platformId, String platformName, String profileImage);
+    User joinOauthUser(BigInteger platformId, String platformName, String profileImage);
 
     Boolean nicknameCheck(String nickname);
